Better RAM detection

+++ include/multiboot.h
@@ -216,8 +216,8 @@ typedef struct multiboot_color multiboot_color_t;
 struct multiboot_mmap_entry
 {
     uint32_t size;
-    uint32_t addr;
-    uint32_t len;
+    uint64_t addr;
+    uint64_t len;
     uint32_t type;
 } __attribute__((packed));
 typedef struct multiboot_mmap_entry multiboot_memory_map_t;
